usda/requirements.txt
requests
beautifulsoup4
selenium
django
django-address
jsonfield
Pillow
django-debug-toolbar
django-extras
